Fixed bugs, add new price model for market, add tests, completed java docs

+++ src/main/java/teamproject/wipeout/game/item/Item.java
@@ -19,7 +19,7 @@ public class Item {
     public String spriteSheetName;
     public String spriteSetName;
     public Integer maxStackSize;
-    public float defaultBuy;
-    public float defaultSell;
+    public double defaultBuy;
+    public double defaultSell;
     
 }
